% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conttime-package.R
\docType{package}
\name{conttime-package}
\alias{conttime}
\alias{conttime-package}
\title{conttime: Time-varying Continuous-time Modeling}
\description{
This package allows for the estimation of time-varying continuous-time models using the Stan framework.
}
\examples{
### Example 1: Running a time-varying continuous-time model with generated data

## install package conttime

# option 1: install version 0.0.65 (2024-08-25)
# 1a. from source:
install.packages( "https://jastag.org/Rpackages/conttime_0.0.65.tar.gz", repos = NULL, type = "source" )

# 1b. Windows binary:
# install.packages( "https://jastag.org/Rpackages/conttime_0.0.65.zip", repos = NULL, type = "win.binary" )

# option 2: install the most recent version from GitHub (note that it may not remain backward compatible in the long term):
# install.packages( "devtools" )
# devtools::install_github( "martinhecht/conttime" )


# load package conttime
library( conttime )

# load example data
data( example1_data )

# generate Stan syntax and starting values
stn <- gen.stan( data.env=data.env, syntax.dir=getwd(), model_name="example1", include.priors=FALSE, chains=1, verbose=TRUE )

# compile the model (this may take approximately 2 minutes)
sm <- stan_model( file=stn$file, model_name=stn$model_name, verbose=FALSE )

# estimate the model (this may take about half a minute)
fit <- optimizing( object=sm, data=stn$data, init=stn$init[[1]], algorithm = "LBFGS", verbose=TRUE )

# get and print results
est <- get.stan( fit=fit, stn=stn, verbose=TRUE )
print( est )

}
