% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_object.R
\docType{data}
\name{data.env}
\alias{data.env}
\title{data.env: Internal Data Environment}
\format{
An object of class \code{environment} of length 75.
}
\usage{
data.env
}
\description{
This object is used internally within the package.
}
\keyword{internal}
