% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_object.R
\docType{data}
\name{design.env}
\alias{design.env}
\title{design.env: Internal Design Environment}
\format{
An object of class \code{environment} of length 11.
}
\usage{
design.env
}
\description{
This object is used internally within the package.
}
\keyword{internal}
