% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.data.R
\name{gen.data}
\alias{gen.data}
\title{Generate data}
\usage{
gen.data(
  design.env,
  seed = "random",
  value.env = NULL,
  gen.data = TRUE,
  tries.max = 100,
  fac = 1.75,
  verbose = TRUE
)
}
\arguments{
\item{design.env}{an environment containing objects created by the \code{gen.design()} function}

\item{seed}{a number used as the seed for \code{set.seed(seed)} or the value "random" for the random generation of a seed}

\item{value.env}{an environment containing the true values of the matrices A0, Achange, and Q0 for data generation; if \code{NULL}, default values are used}

\item{gen.data}{a logical value indicating whether to generate data}

\item{tries.max}{an integer value that specifies the number of attempts to replace persons (via design) with poorly behaved DT matrices}

\item{fac}{a factor used to determine extreme values of matrix properties, where a value is considered extreme if the property exceeds fac times the standard deviation (SD) over time points}

\item{verbose}{a logical value indicating whether to print detailed messages and progress updates during the execution of the function}
}
\value{
An environment is returned containing design characteristics, generated time points, and generated data. Use \code{ls(envir=<returned environment>)} to view its contents.
}
\description{
Generate data for a time-varying continuous-time model based on an individualized longitudinal design.
}
