% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.design.R
\name{gen.design}
\alias{gen.design}
\title{Generate design}
\usage{
gen.design(
  F = 2,
  N = 5,
  T = 3,
  Tdiv = -1,
  deltas = c(0.5, 1, 1.5),
  first.time.points = c(-1, -0.5, 0, 0.5, 1),
  env = NULL,
  seed = "random",
  verbose = FALSE
)
}
\arguments{
\item{F}{number of processes (variables), must be >= 1}

\item{N}{number of persons, must be >= 1}

\item{T}{the maximum number of time points for each person, must be >= 2}

\item{Tdiv}{a negative number or zero indicating the extent to which the individualized number of time points may vary, T - Tdiv must be >= 2}

\item{deltas}{possible time lengths between consecutive time points, used for sampling with the \code{sample()} function}

\item{first.time.points}{possible first time points, used for sampling with the \code{sample()} function}

\item{env}{an environment where return objects are written; if \code{NULL}, a new environment is created}

\item{seed}{a number used as the seed for \code{set.seed(seed)} or the value "random" for the random generation of a seed}

\item{verbose}{a logical value indicating whether to print detailed messages and progress updates during the execution of the function}
}
\value{
An environment is returned containing design characteristics and generated time points. Use \code{ls(envir=<returned environment>)} to view its contents.
}
\description{
Generate an individualized longitudinal design.
}
