% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.stan.R
\name{gen.stan}
\alias{gen.stan}
\title{Generate Stan code}
\usage{
gen.stan(
  data.env,
  syntax.dir = getwd(),
  model.name = "model",
  model.parameters.env = NULL,
  prior.env = NULL,
  start.values.env = NULL,
  KF = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data.env}{an environment containing objects created by the \code{gen.data()} and \code{gen.design()} functions}

\item{syntax.dir}{the directory where the Stan syntax file is written}

\item{model.name}{name of the model, used to name the Stan file as <model.name>.stan}

\item{model.parameters.env}{an environment containing declarations of model parameters in the matrices A0, Achange, and Q0; if \code{NULL}, the default parameterization is used}

\item{prior.env}{an environment containing priors for the elements of the matrices A0, Achange, and Q0; if \code{NULL}, default priors are used}

\item{start.values.env}{an environment containing start values for the elements of the matrices A0, Achange, and Q0; if \code{NULL}, default start values are used}

\item{KF}{use Kalman Filter estimation}

\item{verbose}{a logical value indicating whether to print detailed messages and progress updates during the execution of the function}
}
\value{
A list is returned containing the elements \code{syntax.path}, \code{data}, \code{init}, \code{model.parameters}, \code{par.env}, \code{prior.env}, and \code{start.values.env}.
}
\description{
Generate Stan code and starting values for a time-varying continuous-time model.
}
