% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.stan.R
\name{get.stan}
\alias{get.stan}
\title{Get results}
\usage{
get.stan(
  fit,
  stn,
  true.env = NULL,
  transformed.parameters = FALSE,
  sort.by.bias = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{fit}{the results object from running the \code{stan()} function}

\item{stn}{the results object from running the \code{gen.stan()} function}

\item{true.env}{an environment containing the true values for the model parameters in the matrices A0, Achange, and Q0; used to calculate bias}

\item{transformed.parameters}{a logical value indicating whether the transformed parameters \code{c("lp__", "thetajp", "At", "Qt", "Ajp", "Qjp", "Astarjp", "Qstarjp", "Sigmawjp")} should be included in the returned results, or a character vector specifying which transformed parameters to include}

\item{sort.by.bias}{a logical value indicating whether to sort results by bias (if true values are provided in true.env)}

\item{verbose}{a logical value indicating whether to print detailed messages and progress updates during the execution of the function}
}
\value{
A data frame with parameter estimates (and bias if true values are provided) is returned.
}
\description{
Get results from fitted Stan model
}
