% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.values.R
\name{start.values}
\alias{start.values}
\title{Set starting values}
\usage{
\method{start}{values}(
  data.env,
  chains = 1,
  start.values.env = NULL,
  jitter = TRUE,
  jitter.env = NULL,
  seed = "random",
  par.env = NULL,
  return.init.only = TRUE,
  tries.max = 100,
  fac = 2,
  verbose = TRUE
)
}
\arguments{
\item{data.env}{an environment containing objects created by the \code{gen.data()} and \code{gen.design()} functions}

\item{chains}{number of chains for estimation with Stan}

\item{start.values.env}{an environment containing start values for the elements of the matrices A0, Achange, and Q0; if \code{NULL}, default start values are used}

\item{jitter}{a logical value indicating whether start values should be jittered to provide different start values for each chain}

\item{jitter.env}{an environment containing information on how to jitter the elements of the matrices A0, Achange, and Q0; if \code{NULL}, default jitter routines are used}

\item{seed}{a number used as the seed for \code{set.seed(seed)} or the value "random" for the random generation of a seed}

\item{par.env}{an environment containing declarations of model parameters in the matrices A0, Achange, and Q0. This is used to determine and align whether a parameter is freely estimated (requiring a start value) or fixed (no start value needed). This is optional, as Stan will ignore start values for fixed parameters.}

\item{return.init.only}{a logical value that determines whether only the \code{init} object for function \code{stan} should be returned, or if additional information (such as the seed and starting values in the form of a data frame) should also be included}

\item{tries.max}{an integer value that specifies the number of attempts to find suitable jittered starting values}

\item{fac}{a factor used to determine extreme values of matrix properties, where a value is considered extreme if the property exceeds fac times the standard deviation (SD) over persons and time points}

\item{verbose}{a logical value indicating whether to print detailed messages and progress updates during the execution of the function}
}
\value{
Either the \code{init} object for the \code{stan} function when \code{return.init.only=TRUE}, or a list containing \code{init}, \code{seed.jitter.sv}, and \code{sv} when \code{return.init.only=FALSE}.
}
\description{
Set starting values for estimation with Stan
}
