% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_object.R
\docType{data}
\name{value.env}
\alias{value.env}
\title{value.env: Internal Value Environment}
\format{
An object of class \code{environment} of length 3.
}
\usage{
value.env
}
\description{
This object is used internally within the package.
}
\keyword{internal}
